Complete lesson 06 reusable components refactor
