+++ src/components/TaskItem.jsx
@@ -1,0 +1,9 @@
+function TaskItem({ task }) {
+  return (
+    <li>
+      {task.title} {task.completed ? '✅' : '⏳'}
+    </li>
+  );
+}
+
+export default TaskItem;
