+++ src/components/UserProfile.jsx
@@ -1,0 +1,5 @@
+function UserProfile() {
+  return <h2>Welcome, Student</h2>;
+}
+
+export default UserProfile;
