+++ src/hooks/useTasks.js
@@ -1,0 +1,24 @@
+import { useEffect, useState } from 'react';
+
+function useTasks() {
+  const [tasks, setTasks] = useState([]);
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    const timeout = setTimeout(() => {
+      setTasks([
+        { id: 1, title: 'Learn React', completed: true },
+        { id: 2, title: 'Refactor code', completed: false },
+        { id: 3, title: 'Organize files', completed: false },
+      ]);
+
+      setLoading(false);
+    }, 500);
+
+    return () => clearTimeout(timeout);
+  }, []);
+
+  return { tasks, loading };
+}
+
+export default useTasks;
